% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_eligibility.R
\name{check_eligibility}
\alias{check_eligibility}
\title{Check whether user is eligible to submit to the challenge.}
\usage{
check_eligibility(team_id, owner_id)
}
\arguments{
\item{team_id}{ID (integer string) of the participant's team.}

\item{owner_id}{Synapse user ID (integer string) of the participant.}
}
\value{
If eligible to submit, return `TRUE`; else return `FALSE`.
}
\description{
Check whether user is eligible to submit to the challenge.
}
