% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synapse_session.R
\name{clear_credentials}
\alias{clear_credentials}
\title{Clear any cached Synapse credentials without starting a new R session.}
\usage{
clear_credentials()
}
\value{
None
}
\description{
Clear any cached Synapse credentials without starting a new R session.
}
