% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configure_login.R
\name{configure_login}
\alias{configure_login}
\title{Collect and store Synapse login credentials.}
\usage{
configure_login()
}
\value{
None
}
\description{
Collect and store Synapse login credentials.
}
