% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submit_predictions.R
\name{confirm_submission}
\alias{confirm_submission}
\title{Prompt user to verify whether they want to submit to challenge.}
\usage{
confirm_submission()
}
\description{
Prompt user to verify whether they want to submit to challenge.
}
