% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_submission.R
\name{create_submission}
\alias{create_submission}
\title{Create submission file for RAAD Challenge}
\usage{
create_submission(.data)
}
\arguments{
\item{.data}{A tbl or data frame with participant predictions.}
}
\value{
`create_submission()` returns a string with the file name for the
    locally saved file.
}
\description{
Writes `.data` to CSV file named with user ID and time stamp.
}
