% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synapse_session.R
\name{.api_key_prompt}
\alias{.api_key_prompt}
\title{Collect and store Synapse API key.}
\usage{
.api_key_prompt()
}
\value{
String with user's API key.
}
\description{
Collect and store Synapse API key.
}
