% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_certification.R
\name{.check_certification}
\alias{.check_certification}
\title{Check whether user is certified on Synapse.}
\usage{
.check_certification(owner_id)
}
\arguments{
\item{owner_id}{Synapse user ID (integer string) of the participant.}
}
\value{
If user is certified, return `TRUE`; else return `FALSE`.
}
\description{
Check whether user is certified on Synapse.
}
