% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_eligibility.R
\name{.check_eligibility}
\alias{.check_eligibility}
\title{Check whether user is eligible to submit to the challenge.}
\usage{
.check_eligibility(syn, team_info, owner_id)
}
\arguments{
\item{owner_id}{Synapse user ID (integer string) of the participant.}

\item{team_id}{ID (integer string) of the participant's team.}
}
\value{
If eligible to submit, return `TRUE`; else return `FALSE`.
}
\description{
Check whether user is eligible to submit to the challenge.
}
