% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submit_predictions.R
\name{.confirm_prompt}
\alias{.confirm_prompt}
\title{Prompt user to verify whether they want to submit to challenge.}
\usage{
.confirm_prompt()
}
\value{
None
}
\description{
Prompt user to verify whether they want to submit to challenge.
}
