% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_certification.R
\name{.get_certification_status}
\alias{.get_certification_status}
\title{Retrieve Synapse certification status for user.}
\usage{
.get_certification_status(owner_id)
}
\arguments{
\item{owner_id}{Synapse user ID (integer string) of the participant.}
}
\value{
If user is certified, return `TRUE`; else return `FALSE`.
}
\description{
Retrieve Synapse certification status for user.
}
