% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_eligibility.R
\name{.get_eligibility_data}
\alias{.get_eligibility_data}
\title{Collect eligibility data for a team and its members.}
\usage{
.get_eligibility_data(syn, team_id)
}
\arguments{
\item{team_id}{ID (integer string) of the participant's team.}
}
\description{
Collect eligibility data for a team and its members.
}
