% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_eligibility.R
\name{.get_owner_eligibility}
\alias{.get_owner_eligibility}
\title{Parse eligibility data to check participant elibility.}
\usage{
.get_owner_eligibility(eligibility_data, owner_id)
}
\description{
Parse eligibility data to check participant elibility.
}
