% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_eligibility.R
\name{.get_team_eligibility}
\alias{.get_team_eligibility}
\title{Parse eligibility data to check overall team elibility.}
\usage{
.get_team_eligibility(eligibility_data)
}
\description{
Parse eligibility data to check overall team elibility.
}
