% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synapse_session.R
\name{.lookup_owner_id}
\alias{.lookup_owner_id}
\title{Look up the owner ID for Synapse user.}
\usage{
.lookup_owner_id(syn)
}
\value{
String with Synapse ID for team project.
}
\description{
Look up the owner ID for Synapse user.
}
