% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_team_info.R
\name{.lookup_prediction_folder}
\alias{.lookup_prediction_folder}
\title{Look up folder ID where prediction file is to be stored.}
\usage{
.lookup_prediction_folder(syn, team_name)
}
\arguments{
\item{team_name}{Synapse team name.}
}
\value{
String with Synapse ID for submission folder.
}
\description{
Look up folder ID where prediction file is to be stored.
}
