% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_eligibility.R
\name{.owner_eligibility_msg}
\alias{.owner_eligibility_msg}
\title{Construct message summarizing user submission eligibility.}
\usage{
.owner_eligibility_msg(.data)
}
\arguments{
\item{.data}{A tbl with eligibility status for submitting user.}
}
\value{
String summarizing eligibility status and any reasons that user
    might currently be ineligible to submit.
}
\description{
Construct message summarizing user submission eligibility.
}
