% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submit_predictions.R
\name{.success_msg}
\alias{.success_msg}
\title{Provide user with information about where to find submission results.}
\usage{
.success_msg(filename, entity_id, sub_id)
}
\arguments{
\item{sub_id}{}
}
\description{
Provide user with information about where to find submission results.
}
