% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synapse_session.R
\name{.user_email_prompt}
\alias{.user_email_prompt}
\title{Collect and store registered Synapse email.}
\usage{
.user_email_prompt()
}
\value{
String with user's email address.
}
\description{
Collect and store registered Synapse email.
}
