% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_team_info.R
\name{get_challenge_teams}
\alias{get_challenge_teams}
\title{Collect registered challenge team IDs via Synapse REST API.}
\usage{
get_challenge_teams(challenge_id = "4288")
}
\description{
Collect registered challenge team IDs via Synapse REST API.
}
