% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submit_predictions.R
\name{lookup_owner_id}
\alias{lookup_owner_id}
\title{Look up the owner ID for Synapse user.}
\usage{
lookup_owner_id(user_id, table_id = "syn17091891")
}
\arguments{
\item{user_id}{registered email of the participant.}
}
\value{
String with Synapse ID for team project.
}
\description{
Look up the owner ID for Synapse user.
}
