% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_team_info.R
\name{lookup_prediction_folder}
\alias{lookup_prediction_folder}
\title{Look up folder ID where prediction file is to be stored.}
\usage{
lookup_prediction_folder(project_id)
}
\arguments{
\item{project_id}{Synapse ID of team's submission project.}
}
\value{
String with Synapse ID for submission folder.
}
\description{
Look up folder ID where prediction file is to be stored.
}
