% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_team_info.R
\name{lookup_team_project}
\alias{lookup_team_project}
\title{Look up the project/wiki ID for the specified challenge team.}
\usage{
lookup_team_project(team_id, table_id = "syn17091912")
}
\arguments{
\item{team_id}{ID (integer string) of the participant's team.}
}
\value{
String with Synapse ID for team project.
}
\description{
Look up the project/wiki ID for the specified challenge team.
}
