% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_predictions.R
\name{run_checks}
\alias{run_checks}
\title{Define cases and expectations for formatting checks.}
\usage{
run_checks(.data)
}
\arguments{
\item{.data}{A tbl or data frame with participant predictions.}
}
\value{
None
}
\description{
Define cases and expectations for formatting checks.
}
