% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submit_predictions.R
\name{submit_predictions}
\alias{submit_predictions}
\title{Submit predictions to the RAAD Challenge evaluation queue}
\usage{
submit_predictions(.data, team_id, project_id, validate_only = FALSE)
}
\arguments{
\item{.data}{A tbl or data frame with participant predictions.}

\item{team_id}{ID (7-digit string) of the participant's team.}

\item{project_id}{Synapse ID for the participant's team project, where
submitted predictions will be stored.}

\item{validate_only}{If `TRUE`, check data for any formatting errors but
don't submit to the challenge.}
}
\value{
None
}
\description{
Validates prediction data frame (`.data`) to check for any formatting
errors, saves data to local CSV file, stores file in Synapse, then submits
Synapse entity to RAAD Challenge evaluation queue for scoring.
}
\examples{
# Example prediction data frame
prediction_df <- data.frame(
    Subject = seq_len(100),
    SubPopulation = rep_len(c("Y", "N"), 50)
)

# Submitting predictions for team "1234567" and team project "syn16810564"
submit_predictions(prediction_df, "1234567", "syn16810564")

# To validate data only but not submit:
submit_predictions(prediction_df, "1234567", "syn16810564",
                   validate_only = TRUE)
}
