% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submit_predictions.R
\name{submit_raadc2}
\alias{submit_raadc2}
\title{Submit predictions to the RAAD Challenge evaluation queue.}
\usage{
submit_raadc2(predictions, submitter_id = NULL, validate_only = FALSE,
  skip_validation = FALSE, skip_eligibility_checks = FALSE,
  confirm_submit = TRUE, dry_run = FALSE, syn = GLOBAL_SYN)
}
\arguments{
\item{predictions}{A dataframe/tibble with two columns, \emph{PatientID} and
\emph{RespondingSubgroup}.}

\item{submitter_id}{Participant Synapse registered email. If not provided,
user will be prompted for email before submission.}

\item{validate_only}{If `TRUE`, check data for any formatting errors but
don't submit to the challenge.}

\item{skip_validation}{If `TRUE`, skip formatting checks and submit data
to the challenge.}

\item{dry_run}{If `TRUE`, execute submission steps, but don't store any
data in Synapse.}
}
\value{
None
}
\description{
Validates prediction data frame (`predictions`) to check for any formatting
errors, saves data to local CSV file, stores file in Synapse, then submits
Synapse entity to RAAD Challenge evaluation queue for scoring.
}
\examples{
\dontrun{
# Example prediction data frame
set.seed(2018)
d_predictions <- data.frame(
  PatientID = paste0("Pat",1:400),
  RespondingSubgroup = rep(c("Tecentriq","Chemo"), 200)
)

# Submitting predictions for user "synuser@gene.com"

submit_predictions(d_predictions, submitter_id = "synuser@gene.com")

# To validate data only but not submit:
submit_predictions(d_predictions, submitter_id = "synuser@gene.com",
                   validate_only = TRUE)
# To simulate submission process without uploading or submitting data:
submit_predictions(d_predictions, submitter_id = "synuser@gene.com",
                   dry_run = TRUE)
}
}
