% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synapse_session.R
\name{synapse_login}
\alias{synapse_login}
\title{Convenience wrapper function for Synapse login.}
\usage{
synapse_login(syn, user)
}
\value{
None
}
\description{
Convenience wrapper function for Synapse login.
}
