% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_predictions.R
\name{validate_predictions}
\alias{validate_predictions}
\title{Validate RAADC 2.0 predictions}
\usage{
validate_predictions(predictions)
}
\arguments{
\item{predictions}{A dataframe/tibble with two columns, \emph{PatientID} and
\emph{RespondingSubgroup}.}

\item{subset_patientids}{Vector of patientids, if you only want to include
specific patients in the scoring process. This is used to do the 20% scoring
for the leaderboard.}
}
\value{
If all checks pass, return TRUE; otherwise, raise an error.
}
\description{
\code{validate_predictions} takes a set of predictions and checks for any
formatting errors.
}
\examples{
\dontrun{
set.seed(2018)
patient_nums <- stringr::str_pad(1:1000, width = 5, side = "left", pad = "0")
d_predictions <- data.frame(
  PatientID = stringr::str_c("RAADCV", patient_nums),
  Treatment = rep(c("Tecentriq","Chemo"), 500)
)

validate_predictions(d_predictions)
}
}
