% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_submission.R
\name{create_submission}
\alias{create_submission}
\title{Create submission file for RAAD Challenge.}
\usage{
create_submission(predictions, stamp = FALSE, dry_run = FALSE)
}
\arguments{
\item{predictions}{A dataframe/tibble with two columns, \emph{PatientID} and
\emph{RespondingSubgroup}.}

\item{stamp}{If `TRUE`, name the file according to user and time; otherwise
use simple 'predictions.csv' file name.}

\item{dry_run}{If `TRUE`, return file name but don't create file.}
}
\value{
`create_submission()` returns a string with the file name for the
    locally saved file.
}
\description{
Writes `predictions` to CSV file named with user ID and time stamp.
}
