% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_team_info.R
\name{get_team_info}
\alias{get_team_info}
\title{Collect challenge team information for the specified Synapse user.}
\usage{
get_team_info(syn, owner_id)
}
\arguments{
\item{owner_id}{Synapse user ID (integer string) of the participant.}
}
\description{
Collect challenge team information for the specified Synapse user.
}
