% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_team_info.R
\name{get_team_table}
\alias{get_team_table}
\title{Collect registered challenge team IDs from Synapse table.}
\usage{
get_team_table(table_id = "syn17091912")
}
\description{
Collect registered challenge team IDs from Synapse table.
}
