% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_team_info.R
\name{lookup_team_project}
\alias{lookup_team_project}
\title{Look up the project/wiki ID for the specified challenge team.}
\usage{
lookup_team_project(team_id)
}
\arguments{
\item{team_id}{}
}
\description{
Look up the project/wiki ID for the specified challenge team.
}
