% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_predictions.R
\name{validate_predictions}
\alias{validate_predictions}
\title{Check prediction data for formatting errors.}
\usage{
validate_predictions(.data)
}
\arguments{
\item{.data}{A tbl or data frame with participant predictions.}
}
\value{
None
}
\description{
Validates prediction data frame (`.data`) to check for any formatting
errors.
}
\examples{
# Example prediction data frame
prediction_df <- data.frame(
    Subject = seq_len(100),
    SubPopulation = rep_len(c("Y", "N"), 50)
)

validate_predictions(prediction_df)
}
