% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_predictions.R
\name{validate_predictions}
\alias{validate_predictions}
\title{Validate RAADC 2.0 predictions}
\usage{
validate_predictions(predictions)
}
\arguments{
\item{predictions}{A dataframe/tibble with two columns, \emph{PatientID} and
\emph{RespondingSubgroup}.}

\item{subset_patientids}{Vector of patientids, if you only want to include
specific patients in the scoring process. This is used to do the 20% scoring
for the leaderboard.}
}
\value{
If all checks pass, return TRUE; otherwise, raise an error.
}
\description{
\code{validate_predictions} takes a set of predictions and checks for any
formatting errors.
}
\examples{

paste("Validate example 1")
set.seed(2018)
d_predictions <- data.frame(
  PatientID = paste0("Pat",1:400),
  RespondingSubgroup = rep(c("Tecentriq","Chemo"), 200)
)

validate_predictions(d_predictions)

paste("Validate example 2")

validate_predictions(
  getRAADC2::d_predictions_harbron
)

}
